Add implementation for CTS test for NPU manager.
Creates a test framework for two applications that run NPU inferences,
one in foreground and one in background. The foreground app will begin
running its inferences after the background app, but should finish first
due to higher priority from being in the foreground.

The test is currently written to run on the Edge TPU. It currently fails on Pixel because the Npu Manager logic has not
been implemented yet - this is expected behavior.

This uses NNAPI to run on the TPU but it is deprecated. This should be replaced by a LiteRT
solution which can run on Edge TPU and all vendor NPUs.

Bug: 452691564
Test: atest CtsNpuManagerTestCases (100/100 iterations fail - expected)
Flag: TEST_ONLY
Change-Id: If22aba8330d2c14799a5ed2888f1f086604be759