Refactor priority handling
This adds PriorityManager, which handles the monitoring of app
importance, turning that into a priority, and sending that to the NPU
IScheduling HAL. It also handles "unknown" apps (which do not declare
the android.hardware.npu feature) and sends a priority for them
appropriately, depending on their target SDK.

Apps targeting CINNAMON_BUN will be blocked from NPU access if they do
not declare that they use the NPU in their manifest (`<uses-feature
name="android.hardware.npu"/>`). Currently this is just a warning.

Bug: 467932746
Test: atest CtsNpuManagerTestCases
Flag: com.android.npumanager.npumanager_enabled
Change-Id: I0af103abf740b745b9008e52f36d6c2499390adb